backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
# Core automation library used by real_browser.py
browser-use==0.6.1

